dma: change buffer busy logic
previously buffer would be marked as busy if the fifo had data, even if
the busy bits were marking it as idle

busy = dout or (not empty) means that the only idle state
is dout==0 and empty == 1

This is corrected so that now we will be idle
only when the busy bits are deasserted.

We do not want to switch on FIFO empty because the DMA drains the block
much faster than the DAQ can write it, so it may go empty in the middle
of packets.

The DAQ block writes one burst transfer worth of idle data into the
buffer so the switch should happen then.