fix(mtb): take TIU event ID at the correct time
in some effort in Feb 2023 to reduce latency through the system (commit
1d61dac) I created a 2nd signal in the MTB called the "pre-trigger"
which fires 1 clock cycle earlier than the "trigger".

Basically, pre-trigger fires. This tells the MTB firmware to increment
the event ID by one. But incrementing the ID takes one clock cycle.
Latency could be reduced by 1 clock cycle by using the pre-trigger to
begin some machinery used in sending out the trigger. It seems the logic
going to the TOF digested this pretrigger correctly (recognizing that
the event ID will not come until 1 cycle later). The TIU interface,
however, did not :(

So what was happening is that at the time that the pre-trigger fires,
the TIU trigger line will go high and start serializing the event ID,
but using the *previous* event's ID which was there before the ID
increments in the next clock cycle.

In the new patched firmware: Now, at the time of the pre-trigger, we
immediately pull the trigger line high immediately at the pretrigger for
the precious 10 ns latency reduction but will take the event ID at the
correct time (trigger).